Refactor the Housing Benefit parameter, variable and test files.
