+++ policyengine_uk/parameters/gov/dwp/housing_benefit/allowances/age_threshold/younger.yaml
@@ -1,0 +1,12 @@
+description: A lower Houseing Benefit allowance amount is provided if the oldest adult in the household is under this age threshold.
+metadata:
+  label: Housing Benefit allowance younger age threshold
+  period: year
+  unit: year
+  reference: 
+    - title: The Housing Benefit Regulations 2006 sec. 74
+      href: https://www.legislation.gov.uk/uksi/2006/213/made/data.pdf#page=73
+
+
+values:
+  2019-06-01: 18
