--- policyengine_uk/parameters/gov/dwp/housing_benefit/allowances/child.yaml
@@ -1,14 +1,0 @@
-description: Personal allowance increase for a child for Housing Benefit
-metadata:
-  economy: false
-  label: HB child allowance
-  name: hb_child
-  period: week
-  propagate_metadata_to_children: true
-  reference: https://england.shelter.org.uk/legal/benefits/housing_benefit/benefit_rates/housing_benefit_calculation_rates_2019_to_2020
-  unit: currency-GBP
-  uprating: gov.benefit_uprating_cpi
-values:
-  2015-06-01: 66.9
-  2021-04-01: 68.6
-  2022-04-06: 70.8
