+++ policyengine_uk/parameters/gov/dwp/housing_benefit/allowances/couple/both_under_18.yaml
@@ -2,8 +2,7 @@ description: Personal allowance for Housing Benefit for a couple, both aged unde
   18
 metadata:
   economy: false
-  label: HB personal allowance (couple, under 18)
-  name: hb_couple_u18
+  label: Housing Benefits personal allowance (couple, under 18)
   period: week
   propagate_metadata_to_children: true
   reference: https://england.shelter.org.uk/legal/benefits/housing_benefit/benefit_rates/housing_benefit_calculation_rates_2019_to_2020
