+++ policyengine_uk/parameters/gov/dwp/housing_benefit/means_test/income_disregard/worker_hours.yaml
@@ -1,9 +1,9 @@
 description: Default hours requirement for the Working-Tax-Credit-hours-related element of Housing Benefit
 values:
-  2015-06-01: 30.00
+  2015-06-01: 30
 metadata:
   period: week
   unit: hour
-  label: HB worker element hours requirement
+  label: Housing Benefit worker element hours requirement
   economy: false
   reference: https://www.gov.uk/working-tax-credit
