--- policyengine_uk/parameters/gov/dwp/housing_benefit/means_test/pension_disregard.yaml
@@ -1,10 +1,0 @@
-description: Threshold in occupational and personal pensions, above which Housing
-  Benefit is reduced
-metadata:
-  economy: false
-  period: week
-  label: Housing Benefit pension disregard
-  reference: https://www.iser.essex.ac.uk/files/projects/UKMOD/EUROMOD_country_report.pdf#page=40
-  unit: currency-GBP
-values:
-  2015-06-01: 50.0
