+++ policyengine_uk/parameters/gov/dwp/housing_benefit/non_dep_deduction/age_threshold.yaml
@@ -1,0 +1,12 @@
+description: Disabled child premium for Housing Benefit
+metadata:
+  label: Housing Benefit non dependent deduction age threshold
+  period: year
+  unit: year
+  reference: 
+    - title: The Housing Benefit Regulations 2006 sec. 74
+      href: https://www.legislation.gov.uk/uksi/2006/213/made/data.pdf#page=73
+
+
+values:
+  2019-06-01: 21
