--- policyengine_uk/parameters/gov/dwp/housing_benefit/premiums/disabled_child.yaml
@@ -1,12 +1,0 @@
-description: Disabled child premium for Housing Benefit
-metadata:
-  economy: false
-  label: HB disabled child premium
-  name: hb_dis_child
-  period: week
-  reference: https://england.shelter.org.uk/legal/benefits/housing_benefit/benefit_rates/housing_benefit_calculation_rates_2019_to_2020
-  unit: currency-GBP
-values:
-  2019-06-01: 64.19
-  2021-04-01: 65.94
-  2022-04-06: 68.04
