--- policyengine_uk/parameters/gov/dwp/housing_benefit/premiums/family_lone.yaml
@@ -1,9 +1,0 @@
-description: Family premium for Housing Benefit
-metadata:
-  economy: false
-  period: week
-  propagate_metadata_to_children: true
-  reference: https://england.shelter.org.uk/legal/benefits/housing_benefit/benefit_rates/housing_benefit_calculation_rates_2019_to_2020
-  unit: currency-GBP
-values:
-  2019-06-01: 22.2
