+++ policyengine_uk/variables/gov/dwp/housing_benefit/housing_benefit.py
@@ -1,0 +1,21 @@
+from policyengine_uk.model_api import *
+
+
+class housing_benefit(Variable):
+    label = "Housing Benefit"
+    entity = BenUnit
+    definition_period = YEAR
+    value_type = float
+    unit = GBP
+    defined_for = "would_claim_HB"
+
+    def formula(benunit, period, parameters):
+        housing_benefit_entitlement = benunit(
+            "housing_benefit_pre_benefit_cap", period
+        )
+        benefit_cap_reduction = benunit("benefit_cap_reduction", period)
+        return where(
+            housing_benefit_entitlement > 0,
+            max_(0, housing_benefit_entitlement - benefit_cap_reduction),
+            0,
+        )
