+++ policyengine_uk/variables/gov/dwp/housing_benefit/housing_benefit_reported.py
@@ -1,0 +1,9 @@
+from policyengine_uk.model_api import *
+
+
+class housing_benefit_reported(Variable):
+    value_type = float
+    entity = Person
+    label = "Housing Benefit (reported amount)"
+    definition_period = YEAR
+    unit = GBP
